requirements.txt
Flask==3.1.1
flask-cors==6.0.2
python-dotenv==1.1.0
requests==2.32.3
elevenlabs==1.5.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-generativeai==0.8.3
openai==1.79.0
weasyprint==65.1
python-docx==0.8.10
PyPDF2==4.0.2
pdfplumber==0.11.9
PyMuPDF==1.25.5
beautifulsoup4==4.13.3
pandas==2.2.3
numpy==2.2.3
flask-limiter==3.5.0
flask-caching==2.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
Werkzeug==3.1.3
